--- ilia/src/Components/NavBar/NavBar.tsx
@@ -1,36 +1,0 @@
-import Link from "next/link";
-import React, { ReactNode } from "react";
-
-type NavBarProps = {
-	children: ReactNode;
-};
-function NavBar(props: NavBarProps) {
-	const { children } = props;
-	return (
-		<>
-			<div className="w-screen h-24 bg-blue-300 text-white font-bold flex items-center justify-center">
-				<div className="pl-5 mr-auto text-5xl">Ilia</div>
-				<ul className="flex flex-row w-3/5 justify-center gap-11">
-					<Link href={"/"}>
-						<li>Home</li>
-					</Link>
-					<Link href={"/costumer/createCostumer"}>
-						<li>Create Costumer</li>
-					</Link>
-					<Link href={"/order/createOrder"}>
-						<li>Create Order</li>
-					</Link>
-					<Link href={"/order"}>
-						<li>Orders</li>
-					</Link>
-					<Link href={"/products"}>
-						<li>Products Available</li>
-					</Link>
-				</ul>
-			</div>
-			{children}
-		</>
-	);
-}
-
-export default NavBar;
