+++ ilia/src/pages/costumer/[id].tsx
@@ -1,4 +1,4 @@
-import CostumerProfile from "@/Components/Costumers/CostumerProfile";
+import CostumerProfile from "@/components/Costumers/CostumerProfile";
 import { CostumerContext } from "@/contexts/costumerContext";
 import { Costumer } from "@/models/costumer";
 import { useRouter } from "next/router";
