+++ ilia/src/pages/index.tsx
@@ -1,4 +1,4 @@
-import Costumers from "@/Components/Costumers/Costumers";
+import Costumers from "@/components/Costumers/Costumers";
 
 export default function Home() {
 	return (
