+++ ilia/src/pages/order/createOrder.tsx
@@ -1,4 +1,4 @@
-import OrdersForm from "@/Components/Orders/OrdersForm";
+import OrdersForm from "@/components/Orders/OrdersForm";
 import React from "react";
 
 function createOrder() {
